[1701]: Move the MarkFlagsRequired into flags.go (from helpers.go).
